app/coffee-finder-next/foursquareAPI
Extracted `getNearbyCoffeeStores` into separate file and directory.
