+++ apps/coffee-finder-next/pages/index.tsx
@@ -1,8 +1,8 @@
 import type { NextPage } from "next";
-import { RequestInit } from "next/dist/server/web/spec-extension/request";
 import Head from "next/head";
 import Image from "next/image";
 import { CoffeeStoreCard } from "../components/CoffeeStoreCard";
+import { getNearbyCoffeeStores } from "../hooks/requests/getNearbyCoffeeStores";
 
 const Home: NextPage = (props) => {
     console.dir(props);
@@ -38,32 +38,6 @@ const Home: NextPage = (props) => {
 };
 
 export async function getStaticProps() {
-    async function getNearbyCoffeeStores(latlong: string, limit: number) {
-        try {
-            let options: RequestInit = {
-                method: "GET",
-                headers: {
-                    Accept: "application/json",
-                    Authorization:
-                        typeof process.env.FSQ_PLACES_API_KEY === "string"
-                            ? process.env.FSQ_PLACES_API_KEY
-                            : "",
-                },
-            };
-            let res = await fetch(
-                `https://api.foursquare.com/v3/places/search?query=coffee&ll=${latlong}&limit=${limit.toString()}
-       `,
-                options
-            );
-
-            if (!res) throw new Error("error fetching data");
-            let data = res.json();
-            return data;
-        } catch (error) {
-            console.log(error);
-        }
-    }
-
     let nearbyCoffeeStoresData = await getNearbyCoffeeStores(
         "4.695562523190975,-74.0449747402204",
         6
